refactor(explorer): migrate SELF_KERR to PanelBuilder registry
Phase 2/8 (7th of 9 PlotTypes total).

* New explorer_internals/self_kerr.py hosts
  SelfKerrPanelBuilder plus both absorbed render functions:
  display_qubit_self_kerr (multi-curve, level-pair-selected)
  and display_self_kerr (single-curve oscillator variant).
* Replaces the legacy implicit if self.settings[plot_id]:
  dispatch (whose comment # has settings, so must be qubit-mode
  self-Kerr relied on domain knowledge) with an explicit
  isinstance(subsys, Oscillator) check inside both
  build_panel and build_settings_ui.
* Registry adds PlotType.SELF_KERR; the matching dispatch
  branches are removed from Explorer.build_panel and
  ExplorerSettings.build_settings_ui.

Addresses Batch B review finding L21 from PR #289.
