+++ scqubits/explorer/explorer_panels.py
@@ -95,66 +95,3 @@ def display_cross_kerr(
     )
     axes.axvline(param_slice.param_val, color="gray", linestyle=":")
     return fig, axes
-
-
-@rc_context(matplotlib_settings)
-def display_qubit_self_kerr(
-    sweep: "ParameterSweep",
-    subsys: "QuantumSystem",
-    param_slice: "ParameterSlice",
-    fig_ax: tuple[Figure, Axes],
-    which: list[tuple[int, int]] | None = None,
-) -> tuple[Figure, Axes]:
-    subsys_index = sweep.get_subsys_index(subsys)
-    title = r"Self-Kerr: {}".format(subsys.id_str)
-    ylabel = r"Kerr coefficient $\Lambda^{{{},{}}}_{{ll}}$".format(
-        subsys_index, subsys_index
-    )
-
-    kerr_data = sweep["kerr"][subsys_index, subsys_index]
-    if param_slice.fixed != tuple():
-        kerr_data = kerr_data[param_slice.fixed]
-
-    if not which:
-        level_pairs = list(itertools.combinations(list(range(subsys.truncated_dim)), 2))
-    else:
-        level_pairs = which
-
-    kerr_datasets = []
-    for pair in level_pairs:
-        kerr_datasets.append(kerr_data[..., pair[0], pair[1]])
-    kerr_datasets_arr = np.asarray(kerr_datasets).T
-    kerr_namedarray = NamedSlotsNdarray(kerr_datasets_arr, kerr_data.param_info)
-
-    fig, axes = kerr_namedarray.plot(
-        title=title,
-        label_list=level_pairs,
-        ylabel=ylabel + "[{}]".format(units.get_units()),
-        fig_ax=fig_ax,
-    )
-    axes.axvline(param_slice.param_val, color="gray", linestyle=":")
-    return fig, axes
-
-
-@rc_context(matplotlib_settings)
-def display_self_kerr(
-    sweep: "ParameterSweep",
-    subsys: scqubits.Oscillator | scqubits.KerrOscillator,
-    param_slice: "ParameterSlice",
-    fig_ax: tuple[Figure, Axes],
-) -> tuple[Figure, Axes]:
-    subsys_index = sweep.get_subsys_index(subsys)
-    title = r"Self-Kerr: {}".format(subsys.id_str)
-
-    ylabel = "Kerr coefficient $K_{{{}}}$".format(subsys_index)
-    kerr_data = sweep["kerr"][subsys_index, subsys_index]
-    if param_slice.fixed != tuple():
-        kerr_data = kerr_data[param_slice.fixed]
-
-    fig, axes = kerr_data.plot(
-        title=title,
-        ylabel=ylabel + "[{}]".format(units.get_units()),
-        fig_ax=fig_ax,
-    )
-    axes.axvline(param_slice.param_val, color="gray", linestyle=":")
-    return fig, axes
